Use local context when fetching flags
The context provided with WithContext() during setup of the client
will most likely live for the entire duration of the program, so it
doesn't really do that much when provided to the HTTP client reaching
out to fetch the feature flags.

A better option is to provide a context when needing these functions,
which would allow the caller to cancel them earlier if needed.

Also removed context from functions that weren't using them.